Refactor istio status and workload handlers
