+++ handlers/istio_status.go
@@ -2,22 +2,41 @@ package handlers
 
 import (
 	"net/http"
+
+	"github.com/kiali/kiali/business"
+	"github.com/kiali/kiali/config"
+	"github.com/kiali/kiali/grafana"
+	"github.com/kiali/kiali/istio"
+	"github.com/kiali/kiali/kubernetes"
+	"github.com/kiali/kiali/kubernetes/cache"
+	"github.com/kiali/kiali/prometheus"
+	"github.com/kiali/kiali/tracing"
 )
 
 // IstioStatus returns a list of istio components and its status
-func IstioStatus(w http.ResponseWriter, r *http.Request) {
-	// Get business layer
-	business, err := getBusiness(r)
-	if err != nil {
-		RespondWithError(w, http.StatusInternalServerError, "Services initialization error: "+err.Error())
-		return
-	}
+func IstioStatus(
+	conf *config.Config,
+	kialiCache cache.KialiCache,
+	clientFactory kubernetes.ClientFactory,
+	prom prometheus.ClientInterface,
+	traceClientLoader func() tracing.ClientInterface,
+	discovery istio.MeshDiscovery,
+	cpm business.ControlPlaneMonitor,
+	grafana *grafana.Service,
+) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		business, err := getLayer(r, conf, kialiCache, clientFactory, cpm, prom, traceClientLoader, grafana, discovery)
+		if err != nil {
+			RespondWithError(w, http.StatusInternalServerError, "Services initialization error: "+err.Error())
+			return
+		}
 
-	istioStatus, err := business.IstioStatus.GetStatus(r.Context())
-	if err != nil {
-		handleErrorResponse(w, err)
-		return
-	}
+		istioStatus, err := business.IstioStatus.GetStatus(r.Context())
+		if err != nil {
+			handleErrorResponse(w, err)
+			return
+		}
 
-	RespondWithJSON(w, http.StatusOK, istioStatus)
+		RespondWithJSON(w, http.StatusOK, istioStatus)
+	}
 }
